requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.1
